549 add decimal points for TopSpeed and MinSpeed
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@4522 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: e407cec96046b7cd6bbb8fafadf660cb990e6e86
Former-commit-id: 284a8232c728b40473c099a36faa9ec93267f0e0

+++ src/modules/userinterface/legacymenu/racescreens/raceresultsmenus.cpp
@@ -151,12 +151,12 @@ rmPracticeResults(void *prevHdle, tRmInfo *info, int start)
 		free(str);
 
 		/* Top Spd */
-		snprintf(buf, sizeof(buf), "%d", (int)(GfParmGetNum(results, path, RE_ATTR_TOP_SPEED, NULL, 0) * 3.6));
+		snprintf(buf, sizeof(buf), "%3.1f", (GfParmGetNum(results, path, RE_ATTR_TOP_SPEED, NULL, 0) * 3.6));
 		GfuiMenuCreateLabelControl(rmScrHdle, hmenu, "TopSpeed", true, // From template.
 								   buf, GFUI_TPL_X, y);
 
 		/* Min Spd */
-		snprintf(buf, sizeof(buf), "%d", (int)(GfParmGetNum(results, path, RE_ATTR_BOT_SPEED, NULL, 0) * 3.6));
+		snprintf(buf, sizeof(buf), "%3.1f", (GfParmGetNum(results, path, RE_ATTR_BOT_SPEED, NULL, 0) * 3.6));
 		GfuiMenuCreateLabelControl(rmScrHdle, hmenu, "MinSpeed", true, // From template.
 								   buf, GFUI_TPL_X, y);
 
@@ -308,7 +308,7 @@ rmRaceResults(void *prevHdle, tRmInfo *info, int start)
 								   buf, GFUI_TPL_X, y);
         
         //Top speed
-        snprintf(buf, sizeof(buf), "%d", (int)(GfParmGetNum(results, path, RE_ATTR_TOP_SPEED, NULL, 0) * 3.6));
+        snprintf(buf, sizeof(buf), "%3.1f", (GfParmGetNum(results, path, RE_ATTR_TOP_SPEED, NULL, 0) * 3.6));
 		GfuiMenuCreateLabelControl(rmScrHdle, hmenu, "TopSpeed", true, // From template.
 								   buf, GFUI_TPL_X, y);
